Partial refactoring as preparation for MongoDB tests.
