+++ tests/integration/dbclient/common/src/main/java/io/helidon/tests/integration/dbclient/common/tests/simple/SimpleDeleteIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.tests.integration.dbclient.jdbc.tests;
+package io.helidon.tests.integration.dbclient.common.tests.simple;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -21,12 +21,12 @@ import java.util.concurrent.ExecutionException;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import io.helidon.tests.integration.dbclient.jdbc.AbstractIT;
+import io.helidon.tests.integration.dbclient.common.AbstractIT;
 
-import static io.helidon.tests.integration.dbclient.jdbc.AbstractIT.TYPES;
-import static io.helidon.tests.integration.dbclient.jdbc.AbstractIT.dbClient;
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyDeletePokemon;
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyInsertPokemon;
+import static io.helidon.tests.integration.dbclient.common.AbstractIT.TYPES;
+import static io.helidon.tests.integration.dbclient.common.AbstractIT.dbClient;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyDeletePokemon;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyInsertPokemon;
 
 /**
  * Test set of basic JDBC delete calls.
