+++ tests/integration/dbclient/common/src/main/java/io/helidon/tests/integration/dbclient/common/tests/simple/SimpleGetIT.java
@@ -13,18 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.tests.integration.dbclient.jdbc.tests;
+package io.helidon.tests.integration.dbclient.common.tests.simple;
 
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import org.junit.jupiter.api.Test;
 
 import io.helidon.dbclient.DbRow;
-import io.helidon.tests.integration.dbclient.jdbc.AbstractIT;
+import io.helidon.tests.integration.dbclient.common.AbstractIT;
 
-import static io.helidon.tests.integration.dbclient.jdbc.AbstractIT.POKEMONS;
-import static io.helidon.tests.integration.dbclient.jdbc.AbstractIT.dbClient;
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyPokemon;
+import static io.helidon.tests.integration.dbclient.common.AbstractIT.POKEMONS;
+import static io.helidon.tests.integration.dbclient.common.AbstractIT.dbClient;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyPokemon;
 
 /**
  * Test set of basic JDBC get calls.
