+++ tests/integration/dbclient/common/src/main/java/io/helidon/tests/integration/dbclient/common/tests/simple/SimpleInsertIT.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.tests.integration.dbclient.jdbc.tests;
+package io.helidon.tests.integration.dbclient.common.tests.simple;
 
 import java.util.concurrent.ExecutionException;
 import org.junit.jupiter.api.Test;
 
-import io.helidon.tests.integration.dbclient.jdbc.AbstractIT;
+import io.helidon.tests.integration.dbclient.common.AbstractIT;
 
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyInsertPokemon;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyInsertPokemon;
 
 /**
  * Test set of basic JDBC inserts.
