+++ tests/integration/dbclient/common/src/main/java/io/helidon/tests/integration/dbclient/common/tests/simple/SimpleStatementIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.tests.integration.dbclient.jdbc.tests;
+package io.helidon.tests.integration.dbclient.common.tests.simple;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -23,15 +23,15 @@ import org.junit.jupiter.api.Test;
 
 import io.helidon.dbclient.DbRow;
 import io.helidon.dbclient.DbRows;
-import io.helidon.tests.integration.dbclient.jdbc.AbstractIT;
+import io.helidon.tests.integration.dbclient.common.AbstractIT;
 
-import static io.helidon.tests.integration.dbclient.jdbc.AbstractIT.LAST_POKEMON_ID;
-import static io.helidon.tests.integration.dbclient.jdbc.AbstractIT.TYPES;
-import static io.helidon.tests.integration.dbclient.jdbc.AbstractIT.dbClient;
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyDeletePokemon;
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyInsertPokemon;
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyPokemon;
-import static io.helidon.tests.integration.dbclient.jdbc.tests.Utils.verifyUpdatePokemon;
+import static io.helidon.tests.integration.dbclient.common.AbstractIT.LAST_POKEMON_ID;
+import static io.helidon.tests.integration.dbclient.common.AbstractIT.TYPES;
+import static io.helidon.tests.integration.dbclient.common.AbstractIT.dbClient;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyDeletePokemon;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyInsertPokemon;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyPokemon;
+import static io.helidon.tests.integration.dbclient.common.utils.Utils.verifyUpdatePokemon;
 
 /**
  * Test set of basic JDBC common statement calls.
